scripts/requirements.txt
azure-identity
msgraph-sdk
azure-mgmt-keyvault
azure-mgmt-resource
azure-keyvault-secrets
